Adjust tests to new TextareaHome component
